+++ kernel/sched/ems/ems.h
@@ -35,7 +35,7 @@ static inline int prefer_idle_cpu(struct task_struct *p) { return -1; }
 static inline int group_balancing(struct task_struct *p) { return -1; }
 #endif
 
-extern int task_util(struct task_struct *p);
+extern unsigned long task_util(struct task_struct *p);
 extern int cpu_util_wake(int cpu, struct task_struct *p);
 extern unsigned long task_util_est(struct task_struct *p);
 extern inline unsigned int get_cpu_mips(unsigned int cpu);
